Ajustes no layout

+++ src/app/components/car/car.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
-import { CarService } from 'src/app/services/car.service';
 import { AuthService } from 'src/app/auth/auth.service';
 import { Car } from 'src/app/models/car';
+import { CarService } from 'src/app/services/car.service';
 
 @Component({
   selector: 'app-car',
@@ -10,7 +10,7 @@ import { Car } from 'src/app/models/car';
 })
 export class CarComponent implements OnInit {
   cars: Car[] = [];
-  newCar: Car = { year: undefined, model: '', color: '' };
+  newCar: Car = { year: undefined, model: '', color: ''};
   selectedCarId: number | null = null;
   modalMode: 'create' | 'edit' | 'delete' = 'create';
   showModal: boolean = false;
@@ -30,17 +30,16 @@ export class CarComponent implements OnInit {
 
   logout(): void {
     this.authService.logout();
-    // Redirecionar para a página de login ou outra página
   }
 
   openModal(mode: 'create' | 'edit', carId?: number): void {
     this.modalMode = mode;
     if (mode === 'edit' && carId) {
       const carToEdit = this.cars.find(c => c.id === carId);
-      this.newCar = { ...carToEdit }; // Clonando o carro para edição
+      this.newCar = { ...carToEdit };
       this.selectedCarId = carId;
     } else {
-      this.newCar = { year: undefined, model: '', color: '' }; // Resetando para criação
+      this.newCar = { year: undefined, model: '', color: '' };
       this.selectedCarId = null;
     }
     this.showModal = true;
@@ -56,25 +55,29 @@ export class CarComponent implements OnInit {
     });
   }
 
-  updateCar(carId: number, car: Car): void {
-    this.carService.updateCar(carId, car).subscribe(updatedCar => {
-      if (updatedCar) {
-        const index = this.cars.findIndex(c => c.id === carId);
-        if (index !== -1) {
-          this.cars[index] = updatedCar;
-          console.log('Car updated:', updatedCar);
-          this.closeModal();
+  updateCar(carId: number | null, car: Car): void {
+    if (carId != null || carId != undefined) {
+      this.carService.updateCar(carId, car).subscribe(updatedCar => {
+        if (updatedCar) {
+          const index = this.cars.findIndex(c => c.id === carId);
+          if (index !== -1) {
+            this.cars[index] = updatedCar;
+            console.log('Car updated:', updatedCar);
+            this.closeModal();
+          }
         }
-      }
-    });
+      });
+    }
   }
 
-  deleteCar(carId: number): void {
-    this.carService.deleteCar(carId).subscribe(() => {
-      this.cars = this.cars.filter(c => c.id !== carId);
-      console.log('Car deleted:', carId);
-      this.closeModal();
-    });
+  deleteCar(carId?: number): void {
+    if (carId != null || carId != undefined) {
+      this.carService.deleteCar(carId).subscribe(() => {
+        this.cars = this.cars.filter(c => c.id !== carId);
+        console.log('Car deleted:', carId);
+        this.closeModal();
+      });
+    }
   }
 
   closeModal(): void {
